+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Allowed headings:
 
 ## [Unreleased]
 
+## [3.21.0] - 2022-01-25
+
 ### Added
 
 * `WarningView` has new attribute `defaultPadding` which removes the default padding from the component when set to false. When set to true, or not provided in the xml, component has default `8dp` padding.